August 2015 – September 2015
Developed a Multithreaded application which simulate the Connect 4 game
with 2 Players and 1 referee to decide the winner or the draw. Each
thread gains the access to the game board in a critical section
protected by Mutex and Conditional Variables. Conditional variables were
used to make sure that after each move the access of the game board is
transferred to Referee thread for evaluation. This project was developed
using C language and UNIX System Calls on UNIX platform